Move create pl list to execute (#3 #13 #7 #102)
