+++ includes/pipe.h
@@ -6,7 +6,7 @@
 /*   By: heehkim <heehkim@student.42seoul.kr>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/18 15:50:33 by heehkim           #+#    #+#             */
-/*   Updated: 2022/04/23 19:52:02 by heehkim          ###   ########.fr       */
+/*   Updated: 2022/04/24 15:12:44 by heehkim          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 int		traverse_heredoc(t_ast *ast);
 void	traverse_redirection(t_ast *ast, int *in_fd, int *out_fd);
+int		fork_process(t_data *data);
 
 int		execute(t_data *data);
